requirements.txt
accelerate==0.34.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.0.5
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.5.2
Brotli==1.0.9
bs4==0.0.2
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
cbor==1.0.0
cbor2==5.6.4
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cramjam==2.8.3
crashtest==0.4.1
cycler==0.12.1
dask==2024.8.2
dask-expr==1.1.13
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
distributed==2024.8.2
dulwich==0.21.7
einops==0.8.0
executing==0.8.3
fastapi==0.114.0
fastjsonschema==2.20.0
fastparquet==2024.5.0
filelock==3.13.1
FlagEmbedding==1.2.11
flash_attn==2.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.7
ijson==3.3.0
inscriptis==2.5.0
installer==0.7.0
ipykernel==6.28.0
ipython==8.25.0
ir_datasets==0.5.8
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langdetect==1.0.9
langid==1.1.6
langsmith==0.1.116
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.9
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0
mmh3==4.1.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.6.3
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.3
partd==1.4.2
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
pinecone==5.1.0
pinecone-client==5.0.1
pinecone-notebooks==0.1.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
pip==24.2
pkginfo==1.11.1
platformdirs==3.10.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.43
protobuf==5.28.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.15.1
pyparsing==3.1.4
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==305.1
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
rank-eval==0.1.3
ranx==0.3.20
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.0.1
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.2.0
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trec-car-tools==2.6
trove-classifiers==2024.7.2
types-requests==2.32.0.20240907
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
unlzw3==0.2.2
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.4
voyageai==0.2.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==4.0.2
wcwidth==0.2.5
wget==3.2
wheel==0.43.0
win-inet-pton==1.1.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.10.0
zict==3.0.0
zlib-state==0.1.9
